profile all done

+++ app/Http/Controllers/Admin/UserController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\UserCreateRequest;
 use App\Http\Requests\UserUpdateRequest;
+use App\Models\MoneyTransfer;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\File;
@@ -15,7 +16,7 @@ class UserController extends Controller
 
         $users = User::where('role','admin')->search()->paginate(15);
 
-        return view('admin.users.index',compact('users'));
+        return view('admin.users.admin',compact('users'));
     }
 
 
@@ -23,7 +24,7 @@ class UserController extends Controller
 
         $users = User::where('role','sender')->search()->paginate(15);
 
-        return view('admin.users.index',compact('users'));
+        return view('admin.users.sender',compact('users'));
     }
 
 
@@ -31,7 +32,7 @@ class UserController extends Controller
 
         $users = User::where('role','receiver')->search()->paginate(15);
 
-        return view('admin.users.index',compact('users'));
+        return view('admin.users.receiver',compact('users'));
     }
 
 
@@ -88,6 +89,15 @@ class UserController extends Controller
 
 
     public function delete($id){
+
+        if (auth()->user()->id == $id) {
+            return redirect()->back()->with('error','You can not delete yourself');
+        }
+
+        if (MoneyTransfer::where('sender_id',$id)->exists() || MoneyTransfer::where('receiver_id',$id)->exists()) {
+            return redirect()->back()->with('error','This user have a active or previous payment request.');
+        }
+
         $user = User::find($id);
         if (File::exists(public_path($user->avatar))) {
             File::delete(public_path($user->avatar));
@@ -95,5 +105,25 @@ class UserController extends Controller
         $user->delete();
 
         return redirect()->back()->with('success','User deleted successfully');
+    }
+
+
+    public function recharge(Request $request,$id){
+        
+        $request->validate([
+            'fund' => 'required|numeric|min:1',
+        ]);
+
+
+        $user = User::find($id);
+
+        if ($user->role != 'sender') {
+            return redirect()->back()->with('error','You can not recharge this user');
+        }
+
+        $user->fund = $request->fund + $user->fund;
+        $user->save();
+
+        return redirect()->back()->with('success','User recharge successfully');
     }
 }
